k-attractor experiment (very slow, unstable)

+++ include/display.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <iomanip>
 #include <sstream>
 
@@ -8,7 +9,7 @@ inline std::string display(char const x) {
     if(x >= 32 && x < 127) {
         oss << "'" << x << "'";
     } else {
-        oss << "0x" << std::hex << std::uppercase << (unsigned)x;
+        oss << "0x" << std::hex << std::uppercase << (unsigned int)uint8_t(x);
     }
     return oss.str();
 }
@@ -18,7 +19,7 @@ inline std::string display_inline(char const x) {
     if(x >= 32 && x < 127) {
         oss << x;
     } else {
-        oss << "<0x" << std::hex << std::uppercase << (unsigned)x << ">";
+        oss << "<0x" << std::hex << std::uppercase << (unsigned int)uint8_t(x) << ">";
     }
     return oss.str();
 }
